onegt-release-OneGT_MVP_0.1/backend/requirements.txt
# Backend dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
gspread==6.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
pydantic==2.5.3
python-multipart==0.0.6
cachetools==5.3.2
python-jose[cryptography]==3.3.0
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
aiosmtplib==3.0.1
